xoohoox-backend/xoohoox-backend/requirements.txt
PyYAML==6.0.1
python-dotenv==1.0.0
requests==2.26.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.1
pydantic==2.5.2
python-decouple==3.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
redis==5.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.1.0.post1
aiofiles==0.7.0
python-dateutil==2.8.2
pytz==2021.1 